Add Dockerized Metalium Installation
The user will be prompted about installing Docker and Metalium.
If they want to install Metalium, we will install Docker and create
a script `tt-metalium` in `~/.local/bin` which gives access to a
docker-compose.yml in `~/.tt-metalium`. Then we pull the latest
metalium image. The user can then run `tt-metalium` to get access
to the image or run commands within it.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>