Handle missing jq when script starts
If jq is missing when the script starts, we handle the error in the
scripts which rely on it to fetch the TT software versions. If the user
provides values or the functions succeed, we can print those versions at
the beginning of the script, which is preferred. Otherwise, we can retry
after installing the base packages, which include jq. If we still can't
get values, then throw and error and exit.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>