Replace usage of wget with curl
Most distributions ship with curl already installed but wget is often
not included with the base packages.

Removing the use of curl reduces the number of dependencies needed for
the script to function properly.

Furthermore, the installer already suggests using curl to fetch the
script itself so it makes sense to re-use curl as much as possible.