Flipping default on systemd install checking
I came at the systemd checking the way I normally do in bash which is
looking for more explicit activations vs. relying on the backwards way
bash does true/false.  Somewhere in this, I think during a review, we
flipped it back to 0 as the default and I forgot I was doing it the
other way around.

Flipping the default back to "not negative" so to speak, and this sorts
the default case the way we are expecting.

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>