Dropping 20.04 testing for support
Ubuntu 20.04 is already well past EOL, and now that we've moved towards
providing wheels on pypi for things we aren't compiling the pieces on an
end system if we don't have to.

For a variety of reasons pyluwen now
basically doesn't support the older python that's on 20.04 without
direct compilation.  Dropping this as a result.

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>