Finalize mode-repository-beta
For now, this is super restrictive, because I want everyone to have the
same experience (installing all the packages) which should both reduce
user confusion and give us a better idea of where each package fails.
For now, this flag disables the existing dkms, hugepages, and sfpi
install functions, then installs the repos and the software from them.

Signed-off-by: June Knauth (tenstorrent) <jknauth@tenstorrent.com>