Reverting part of 95eaa4e
95eaa4e5ef9b2b495aa09b47fe0f0dd4b4c34a36
or
https://github.com/tenstorrent/tt-installer/commit/95eaa4e5ef9b2b495aa09b47fe0f0dd4b4c34a36

adjusts some lints from shellcheck, except that it doesn't confirm how
it's supposed to work, in this case with pipx and such it wraps things
in "'s that should be interpretted and causes failures that aren't
intended.

This reverts that, adds an appropriate shellcheck dodge and adds a rough
explanation for 'why' that shellcheck dodge needs to be there.  This
should resolve several issues I've seen and/or seen reported,
particulalry with system pre-install scripting and using pipx

Signed-off-by: John 'Warthog9' Hawley <jhawley@tenstorrent.com>