Add "install-inference-server" function
This is implemented according to our standard methodology. It simply
clones tt-inference-server to ~/.local/lib and creates a small wrapper
script in ~/.local/bin to allow the user to access run.py directly. Our
next steps should be to move boilerplate out of the inference server
setup process and into installer, or better yet get rid of it
altogether.

Signed-off-by: June Knauth (tt) <jknauth@tenstorrent.com>